test/fixtures/wpt/wasm/webapi/esm-integration/worker-import-source-phase.tentative.html

<!DOCTYPE html>
<title>Testing import of WebAssembly source phase from JavaScript worker</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });
const worker = new Worker("resources/worker-source-phase.js", { type: "module" });
worker.onmessage = (msg) => {
  assert_equals(msg.data, 42);
  done();
}
</script>
